1) Made sure that each file object has a static init    and atExit function that is automatically called    if this file object class is specified in at least    one file pattern. 2) Started to improve message class to capture type    conflicts at compile time.
git-svn-id: https://svn.code.sf.net/p/alio/code/trunk@32 c7bcb0b3-e99e-4bc6-9bfa-5fb3e3c0becf

+++ tools/message.cpp
@@ -78,13 +78,13 @@ void Message::clear()
 /** Reserves the memory for a message. 
  *  \param size Number of bytes to reserve.
  */
-void Message::allocate(int size)
+void Message::allocate(size_t size)
 {
     m_data_size = size+1;
     m_data      = new char[size+1];
     if(!m_data)
     {
-        printf("can't allocate '%d' bytes.\n", size+1);
+        printf("can't allocate '%ld' bytes.\n", size+1);
         assert(false);
     }
     m_data[0]   = m_type;
